+++ backend/settings.gradle
@@ -2,5 +2,4 @@ rootProject.name = 'backend'
 include 'techpick-batch'
 include 'techpick-core'
 include 'techpick-api'
-include 'techpick-security'
