+++ backend/techpick-api/build.gradle
@@ -11,7 +11,14 @@ repositories {
 
 dependencies {
     implementation project(":techpick-core")
-    implementation project(":techpick-security")
+
+    // spring security and oauth client
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+    testImplementation 'org.springframework.security:spring-security-test'
+
+    // jwt
+    implementation 'io.jsonwebtoken:jjwt:0.9.1'
 
 }
