+++ backend/techpick-api/src/test/resources/application-security.yaml
