--- backend/techpick-security/build.gradle
@@ -1,34 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'kernel360'
-version = '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation project(":techpick-core")
-
-    // spring security and oauth client
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
-    testImplementation 'org.springframework.security:spring-security-test'
-
-    // jwt
-    implementation 'io.jsonwebtoken:jjwt:0.9.1'
-}
-
-tasks.named('test') {
-    enabled = false
-}
-
-bootJar {
-    enabled = false
-}
-
-jar {
-    enabled = true
-}
